move raster conversion funcs to new module
